PERF: Use smaller (16x16) images in KullbackLeibler metric test
The `KullbackLeiblerCompareHistogramImageToImageMetric` test appeared to be the
most time consuming test of ITK's test framework, taking more than 9 seconds on
Windows, on GitHub Actions. Originally it had test images of 100 x 100 pixels, with
commit de0f47aee76c3ee69504c5663ba284e32ba03e44 "Tests for the CompareHistogramImageToImageMetric
class and the KullbackLieblerCompareHistogramImageToImageMetric class", Luis Ibanez, 29 Dec 2003.
However, the same code coverage may be achieved with images of 16 x 16.