STYLE: Use `{}` instead of dummy as `NumericTraits::GetLength` argument
It appears unnecessary to use a local variable as function argument of
`NumericTraits::GetLength(const T &)`, when this local variable is a "dummy",
that is only just default-initialized, or initialized by `T()` or `{}`.

This commit removes such local "dummy" variables, and passes `{}` to those
`NumericTraits::GetLength` calls that had such a dummy as argument.

Follow-up to pull request https://github.com/InsightSoftwareConsortium/ITK/pull/3710
commit be8cb4233b6510f436333213b7cac7c9ab7c07a7
"STYLE: Use `{}` as `MeasurementVectorTraits::IsResizable` argument"