STYLE: Remove backslash + indent from literals, to avoid unwanted spaces
When a backslash at the end of a line of code was used as continuation character
_inside_ a string literal, an indentation at the next line would unintentionally
become part of the string itself.

This commit has removed the backslash, the line break, and the indentation from
those string literals. Clang-Format has taken care of splitting the literals
over multiple lines properly.

It affects a QuadEdgeMeshEulerOperatorDeleteCenterVertexFunction debug message,
an exception message of Versor, and various exception messages from
"Numerics/Statistics".