COMP: Fix MSVC warning C26495: `ZwQuerySystemInformation` uninitialized
Fixed VS2019 Code Analysis warnings, saying:
> warning C26495: Variable 'itk::WindowsMemoryUsageObserver::ZwQuerySystemInformation' is uninitialized.
> Always initialize a member variable (type.6).

Also added an in-class default member initializer to m_hNTLib, which was not yet
initialized by the default-constructor of WindowsMemoryUsageObserver either.

Note that `WindowsMemoryUsageObserver` is a rather "expensive" class
(having a virtual table) and its default-constructor is already non-trivial,
so the extra cost of the added initialization appears relatively low.