BUG: Set ivar to appropriate value when condition is met
Set the appropriate value to the `itk::OnePlusOneEvolutionaryOptimizer`
`m_Initialized` ivar when the condition described in the documentation
for its purpose is fulfilled.

The variable was initialized to `false` in the constructor and its value
was not being changed anywhere else, and the variable is not meant to be
changed through a setter method.