COMP: Decide early whether to use -msse2
Before this change, the -msse2 flag would be provided to the g++ compiler across
the whole project if the CMake cache variable VNL_CONFIG_ENABLE_SSE2_ROUNDING is
1/ON/YES/etc. Usually VNL_CONFIG_ENABLE_SSE2_ROUNDING would be empty in the
first CMake run. If later during that CMake run, VXL_HAS_SSE2_HARDWARE_SUPPORT
is determined to be 1, VNL_CONFIG_ENABLE_SSE2_ROUNDING would be 1 in any
subsequent CMake run, triggering a lot of recompilations with a different set of
compiler flags.

This commit adds a modified version of the check from VXL into VTK, so it runs
early enough.