COMP: Move RayCastHelper from unnamed namespace to private class section
Put the RayCastInterpolateImageFunction helper class that is not meant to be
exposed to the user inside RayCastInterpolateImageFunction, as a private nested
type, instead of in an unnamed namespace. Header files (including *.hxx files)
just shouldn't use an unnamed namespace.

Avoids a Clang-Tidy 17.0.0 warning saying:
> warning G253035D5: do not use unnamed namespaces in header files [google-build-namespaces]

Following C++ Core Guidelines, September 23, 2022 "Don’t use an unnamed
(anonymous) namespace in a header",
http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#sf21-dont-use-an-unnamed-anonymous-namespace-in-a-header