STYLE: Remove std::allocator from TestImportImageContainer (Common/test)
Replaced the use of `std::allocator` by the equivalent `new Element[]` and
`delete[]` statements. Renamed `m_MemoryAllocatedByAllocator` to
`m_MemoryAllocatedByTestImportImageContainer`.

Aims to simplify the code, and make it easier to upgrade to C++17 and further.
C++17 has deprecated the `std::allocator::allocate` and
`std::allocator::destroy` function calls that are removed with this commit.