ENH: Add C++17 `[[nodiscard]]` to Image "Transform" member functions
It would usually be a mistake when a user would unintentionally ignore the
return value of those member functions. Specifically, the overloads of
`TransformPhysicalPointToIndex` and `TransformPhysicalPointToContinuousIndex`
that have two parameters both return a `bool` value. If this return value can be
ignored, it is preferable (for performance reasons) to call the corresponding
overload that has one parameter instead.

The compiler may now produce a warning when the return value of any of those
member functions is unintentionally ignored.