STYLE: Replace `sizeof(array) / sizeof(array[0])` with C++17 `std::size`
By Visual Studio, Replace in Files, using the following regular expression:

    Find what: sizeof\((\w+)\) / sizeof\(\1\[0\]\)
    Find what: sizeof\((.+)\) / sizeof\(\*\1\)
    Find what: sizeof\((\w+)\) / sizeof\(.*\)
    Replace with: std::size($1)

(The third regular expression, `sizeof\((\w+)\) / sizeof\(.*\)` was only
applied to itkConvertBufferTest.cxx, as it cannot be generally applied).

`std::size(array)` appears to be a more readable and safer way to retrieve the
number of elements of a C-style array. It is also `constexpr`.