ENH: Add series_uid arg to imread() for dicom series selection
If filename is a directory of DICOM files, and series_uid is a string, it will read the DICOM series that matches this series_uid name.  If series_uid is an integer, it will read the series_uid'th DICOM series in the directory (using python notation, e.g., -1 is the last DICOM series in the directory). If unspecified, it will read the first series in the directory.

This does represent a change in the default behavior in that it used to
raise an exception if a DICOM directory with more than one series was
passed to imread, and now it will read the first series in that
directory.