ENH: Use generator/lambda ITK filter for pixel computation
Improvements to the LaplacianSharpeningImageFilter.

Replaced a sequence of math filter with a lambda function used by a
binary generator image filter. Additionally, replace the intensity
windowing filter with the original math computed in a unary generator
using a lambda function.

Additionally, a non-filter calculator was replaced with the
MinimumMaximumImageFilter. Filters are deleted by setting the pointer
to nullptr after usage.