DOC: Document miscellaneous cell class methods in header files
Document miscellaneous cell class methods in header files: Doxygen picks
the method documentation from the header files, and leaves the method
documentation empty if it does not find it there.

Increase consistency in the characters used for documentation blocks
and follow the ITK SW Guideline: prefer using single-line comments for
separation of logical blocks.

Take advantage of the commit to fix typos in the method documentation.