DOC: Document `RGBGibbsPriorFilter` class methods in header file only
Document `itk::RGBGibbsPriorFilter` class methods in header file only:
Doxygen picks the method documentation from the header files, and leaves
the method documentation empty if it does not find it there. So remove
the method documentation from the implementation file.

Take advantage of the commit to use the Doxygen math syntax where math
symbols are used in the doumented methods.