STYLE: Prefer using `constexpr <type>` over `struct` for image dimension
Prefer declaring image dimensions as `constexpr <type>` over `struct` in
tests: a struct does not provide any particular advantage to host the
image dimensions.

Declare the type as an `unsigned int` and change the loop index
variables accordingly to avoid signed/unsiged comparison mismatch
warnings. Take advantage of the commit to limit the loop index
variables' scope to the loop.