DOC: Add Sphinx Documentation rendering
This adds Myst (markdown) Sphinx documentation. This is to be rendered on
ReadTheDocs and deployed on docs.itk.org. A start for issue #4063.
Changes and more content are expected in the future! In particular, the
old wiki content, Software Guide content, and Tutorial. It
uses the clean and elegant Furo theme, which works with Myst and is also
used by docs.vtk.org and wasm.itk.org.

This incorporates the markdown files in the Documentation/ directory. It
renders a clean build without warnings. A few changes were required for
to avoid warknings, e.g. header specifications and cross-link
specifications.

Symlinks were created at the top level for CODE_OF_CONDUCT.md and
CONTRIBUTING.md, but these should be replace by real files with links to
the rendered content once published for GitHub access.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>