PERF: Replace std::function w/ template argument ParallelizeImageRegion
Declared the `funcP` parameter of `MultiThreaderBase` template member functions
`ParallelizeImageRegion` and `ParallelizeImageRegionRestrictDirection` by a
template argument, `TFunction`, instead of by `TemplatedThreadingFunctorType`.
The `funcP` parameter is typically specified by a lambda. Using `TFunction` is
more lightweight, as it avoids converting the lambda to an
`TemplatedThreadingFunctorType`. (`TemplatedThreadingFunctorType` is an alias of
`std::function`.)