STYLE: Remove itkSetObjectMacro if itkSetConstObjectMacro is also there
It is not useful to have both `itkSetObjectMacro(name, type)` and
`itkSetConstObjectMacro(name, type)` with the very same `name` and `type`, in
one and the same class.

Replaced the `itkSetObjectMacro` calls with the corresponding Set member
functions, implemented by simply calling the "const" overload (as defined by the
corresponding `itkSetConstObjectMacro` calls). Intended to remove those member
functions in the future.

Found by regular expression:

    ^([ ]*)itkSetObjectMacro\((.+)\);\r\n\1itkSetConstObjectMacro\(\2\);