ENH: Let `x::New()` initialize the created object by doing `new x()`
Replace `new x` with `new x()`, in the definitions of `itkSimpleNewMacro` and
`itkFactorylessNewMacro`. This ensures that objects created by `x::New()` are
"value-initialized". In practice, this will zero-initialize data members that
might otherwise be left uninitialized.

Added GTests to check that `x::New()` does indeed properly initialize the
created object.