BUG: SetIdentity must retain FixedParameters
When setting a transform to represent an Identity
mapping, it is crucial that the stationary component
of the transform is preserved.

```python
tfm.SetCenter( [1,2,3] );
tfm.SetIdentity();
out=tfm.GetCenter();
assert( out == [1,2,3] );
```

===============

Add test demonstrating change in fixed paramers

When applying SetIdentity(), the fixed parameters should not be modified.
The fixed paramers represent the stationary component of the transform
and setting the transform back to an identity should not modify the
stationary components.