ENH: Handle NIFTI space and time units more consistently
BUG: xyzt_units set incorrectly in output.

Previously, the correct xyzt code was assigned to the 'xyz_units' variable,
but time gets masked out because it's only for space units, resulting in
xyzt_units=2.

Fix by setting 'time_units' to NIFTI_UNITS_SEC.

This produces images with xyzt_units == 10, which is correct for ITK output
(NIFTI_UNITS_MM | NIFTI_UNITS_SEC)

ENH: Convert the origin to mm and s.

Previously, the spacing of the first three dimensions were converted to mm,
but the fourth dimension (time) was not converted to s.

Also, the origins of space and time did not change units on read.

ENH: Use the toffset field to encode the origin in the fourth dimension.