BUG: Provide an API for SpatialObjectProperties that works with python
The dictionary used by SpatialObjectProperties is queried using a function
that returns the dictionary value in a variable that is passed by reference:
https://itk.org/Doxygen/html/classitk_1_1SpatialObjectProperty.html#aff2b1c659a8e77bdb6afeef0bab037fb

bool itk::SpatialObjectProperty::GetTagScalarValue ( const std::string & tag,
double & value
) const

This function is not correctly wrapped by SWIG, and cannot be used from within python.

Here is a snippet of code that demonstrates the issue:

In [1]: import itk
In [2]: so = itk.TubeSpatialObject[3].New()
In [3]: so.GetProperty().SetTagScalarValue("foo",1)
In [4]: ret = 0.0
In [5]: so.GetProperty().GetTagScalarValue("foo",ret)
---------------------------------------------------------------------------
TypeError                                 Traceback (most recent call last)
Cell In[5], line 1
----> 1 so.GetProperty().GetTagScalarValue("foo",ret)
TypeError: in method 'itkSpatialObjectProperty_GetTagScalarValue', argument 3 of type 'double &'

This commit adds new functions that can be correctly wrapped.  Old
functions are unmodified to preserve backward compatibility.