ENH: Optionally limit an ImageMask to a specific pixel value
Nominally an ImageMaskSpatialObject considers any non-zero pixel
to indicate that the object is present at that point.

With this change, it is possible to specify a MaskValue and to
enable the use of that MaskValue (via the UseMaskValue functions)
so that a specific pixel value is used to indicate the presence
of that object.

This is particularly useful when a single mask image defined multiple
objects via different values.   That single image can be assigned
to multiple ImageMaskSpatialObjects, each using a different MaskValue.
- reduces memory requirements, etc.

A new set of tests were added to confirm proper function.

Default behaviour is unchanged.