COMP: Fixed unused legacy remove Macros
This is a commit to solve the Wunused-macros warning generated by clang related to ITK_LEGACY_REMOVE macro.

The change is to create an ifnded macro with it even if ITK_LEGACY_REMOVE is not used in the file. In itkConfigue.h, ITK_LEGACY_REMOVE is defined there if the code is in the legacy remove scenario or not. We need to keep ITK_LEGACY_REMOVE to create a flag that the code is legacy.