COMP: Fix unused macros related to MSVC and OS
This is to fix the -Wunused-macros generated by clang for macros related to MSVC and Microsoft.

For itkNumericTraitsTest.cxx, the _MSC_VER is added to respect the comment.

For itkTestDriver.cxx, ITK_TEST_DRIVER_PATH_SLASH / is related to non Windows machine and there is none in the code and hence it is removed. Now, for ITK_TEST_DRIVER_PATH_SLASH \\, it is not found in the entire code. This is the reason why it is removed.