COMP: Fix -Wextra-semi Warnings for Macro
This is to see if the following changes regarding the -Wextra-semi (semi = semicolon) warning of clang can be done throughout the repository.

This is for a macro that is not function-like. It should not contain a semicolon. It is a method that is declared and defined in a header so no semi colon is needed.

Are these following changes accepted to proceed with the other cases in the repository?