DOC: Start ITK 6 migration guide
This document provides information for developers on how to migrate from
ITK 5 to ITK 6. It contains information on what backwards-incompatible
changes were made, such as the removal of deprecated code, why they made,
and how to migrate old code no the new code.

Developers making backwards-incompatible changes for ITK 6 should add
entries to this document as the changes are made.

The ITK 6 migration guide is similar to the ITK 5 migration guide with the
following modifications:

- The current and previous migration guide are now developed on GitHub
  and published to the ITK Sphinx Documentation.
- We now have the consistent and simpler ITK_LEGACY_REMOVE and
  ITK_FUTURE_LEGACY_REMOVE options without the additional
  *_COMPATIBILITY option.
- We have "most" instead of "all" ITK_LEGACY_REMOVE,
  ITK_FUTURE_LEGACY_REMOVE to allow for exceptions on an as-needed basis
  to prevent forcing deprecations that may be premature in terms of
  their removal-cost-benefit or the availability of migration
  options.