COMP: Fix python wrapping build state to match python version
If initially configured with python >= 3.11 before selecting a
python version <3.11 as the python environment, then the
state of the ITK(USE_PYTHON_LIMITED_API was incorrecly
cached to retain the compiler settings desired for
building with the original 3.11 python version.

Py_buffer is only included as part of the limited
python API in 3.11 or greater versions.

Compilation would fail due to building with
-DPy_LIMITED_API to restrict the python interfaces
available.