BUG: Python wrapping needs std::string pass by value in some cases
std::string * doesn't have a desctructor in the SWIG wrapping used
by ITK.  As a result, spam messages are being written when
std::string is passed by reference or returned by reference in
certain cases.  Also, return by references are remaining as
abstract SWIG Objects that are not easy to convert to standard
python types.

This commit fixes this issue for SpatialObjectProperty member
functions.