ENH: Add DICOMOrientation class from remote module
The class provides anatomical orientation based enums for describing
coordinate systems. In ITK the physical coordinate systems is the same
as the DICOM standard. DICOM has the following specification:

If Anatomical Orientation Type (0010,2210) is absent or has a value of
BIPED, the x-axis is increasing to the left hand side of the
patient. The y-axis is increasing to the posterior side of the
patient. The z-axis is increasing toward the head of the patient.

Which describes the direction of increasing. The above coordidate
system is described as "LPS" with this enumeration and corresponds to
an ITK image with an identity direction cosine matrix.