COMP: change itkLegacyMacro to never require a trailing semi colon
Previously, some places *required* a trailing semi and some places required *no* trailing semi.

For example, I see no way to force this kind of usage to have a trailing semi, because that would stick a semi right before the function definition, which isn't allowed:

```
  itkLegacyMacro(static ThreadIdType GetGlobalDefaultNumberOfThreads())
  {
                          "can affect all MultiThreaderBase's derived classes in ITK");
    return Superclass::GetGlobalDefaultNumberOfThreads();
  }
```

Such existing uses only work if inside `#if !defined(ITK_LEGACY_REMOVE)` because only that variation of the definition of the `itkLegacyMacro` macro doesn't require a trailing semi.

It's rather strange for users to both require and forbid traling semis, so this change chooses the only way that can work in all cases.

Fixes various -Wextra-semi warnings.