BUG: SwapWriteRange functions should just write, when `sizeof(T)` is 1
When `sizeof(T) == 1`, the `ByteSwapper<T>` member functions
`SwapWriteRangeFromSystemToBigEndian` and `SwapWriteRangeFromSystemToLittleEndian`
should just write the specified range to the output stream. (In that case,
byteswapping is not necessary.)

The original code did nothing, in that case. It did not swap, but it did not
write the range either.

Fixed by this commit.