ENH: Use pixi to install pre-commit, python for pre-commit
On Windows, Python is not available by default. The
SetupForDevelopment.sh script does not complete successfully in a
default build on Windows.

This uses pixi to install pre-commit, python for pre-commit, so the
contribution experience works out-of-the-box for experienced and new
contributors to ITK.

This also improves the isolation and maintainability of pre-commit
tooling. Everything needed for a commit is local to the ITK repository.
We also get more reproducible versions of Python and pre-commit via
pixi's configuration in pyproject.toml and pixi.lock.

pixi is a self-contained, single-file, cross-platform Rust binary. A local version is
installed to .git/hooks/pixi/bin/pixi. Pixi's python and pre-commit is
installed to .pixi/.

The installation of tools is:

  SetupForDevelopment.sh -> curl -> pixi -> python, pre-commit -> pre-commit's hook dependencies

This also makes it easy to run the pre-commit hooks on demand on all
files via:

```
pixi run pre-commit-run
```