ENH: Shaped NumPy array input to SetPointsByCoordinates
Accept and n_points by dimension shaped NumPy array input to
GetPointsByCoordinates, which is the expected shape for point
coordinates. Continue to also support flat array inputs.

Also add a corresponding shaped NumPy output in a GetPointsByCoordinates method.

Add GetPointDimension() to PointSetBase. This provides access to the
point dimension at runtime for GetPointsByCoordinates and is similar to
GetImageDimension() on ImageBase.