STYLE: clang-tidy bugprone-implicit-widening-of-multiplication-result
The check diagnoses instances where a result of a multiplication is implicitly
widened, and suggests (with fix-it) to either silence the code by making
widening explicit, or to perform the multiplication in a wider type, to avoid
the widening afterwards.

This is fine in general, but if width * height overflows, you end up
wrapping back to the beginning of base instead of processing the entire
requested buffer.

Indeed, this only matters for pretty large buffers (4GB+), but that can
happen very easily for example in image processing, where for that to
happen you “only” need a ~269MPix image.