COMP: Set compile time if statements constexpr
If constexpr feature was introduced to allow compile-time branching
based on constant expressions. Unlike regular if statements, which are
evaluated at runtime, the if constexpr allows the compiler to discard
branches of code that do not apply. It means only the branch of code for
which the condition is true is compiled and other code branch is
discarded during compilation.