STYLE: Use auto for variable type matches
When the type of the initializer expression matches the declaration
use 'auto' to reduce the verbosity (redundnacy).

This check is responsible for using the auto type specifier for variable
declarations to improve code readability and maintainability.

The auto type specifier will only be introduced in situations where the
variable type matches the type of the initializer expression. In other words
auto should deduce the same type that was originally spelled in the source