STYLE: Use range-based loops from C++11
C++11 Range based for loops can be used instead of accessing
container via indexed random access operators.

Used as a more readable equivalent to the traditional for loop operating
over a range of values, such as all elements in a container, in the
forward direction..

Range based loopes are more explicit for only computing the
end location once for containers.