BUG: Set##membervar needs call Modified()
When the state of an object changes through a
Set##membervar( T value) function, the object
must have this->Modified() called to indicate
that it needs to be re-run.

A few Set##membervar functions were missing
the this->Modified() call.

There was a lot of code that was the same
across many of these Set##membervar functions.
The common code was converted to a template
function to remove redundancy.

The template function variant of the code
was simplified to perform the element-wise
sameness check, setting a flag to indicate
a change of state, and return true/false
to indicate state change allows for reduced
code redundancy.