ENH: Exploring the behavior of SetSpacing
The SpatialObjectToImageFilter::SetSpacing has a specialized behavior
when zeros are passed for the spacing size.

It is unclear what the intended behavior is.

Expanded the test suite to demonstrate that sending all zeros as the
spacing values to update results in no changes to the state of
itkSpatialObjectToImageFilter, but does trigger a Modified status on the
Object.