PERF: Prefer default initialization for std::string
Return {} may be the fastest, because it just calls the
default-constructor of std::string.  (return { "" } and return "" may be
slightly slower because they convert a literal string to an std::string.

This probably does not have a substantive impact on the current
uses, but implementing to keep the style consistent.