BUG: fixed duplicated if-else condition, enum was copy-pasted
Looking at the enum, I think the other value was meant to be used, and this was copy-paste.

Probably was never noticed because it's only in PrintSelf().

cppcheck warning was:

multiCondition,/ITK/Modules/Filtering/DiffusionTensorImage/include/itkDiffusionTensor3DReconstructionImageFilter.hxx:631,style,Expression is always false because 'else if' condition matches previous condition at line 626.