DOC: Update maintenance branch for 5.4 series
Someone deleted the `5.4` branch from the repository. And it had not
been updated in the last seven months, but a `release-5.4` branch was
created and has been updated. In prior section, our documentation
references `release-5.4` as the maintenance branch.

We had anticipated using the `5.4` to present stable builds of the
documentation in ReadTheDocs -- their versioning recognizes `5.4` but
not `release-5.4`. However, we can also build versions of the tags and
present that in ReadTheDocs. And, this allows use to used tagged
subproject, like `itkdoxygen`, and RTD will provide unified search over
the tags. With the tags, and the `release` builds, the `X.X` RTD
versions are redundant.