BUG: Consistency of levels and iterations
If the number of levels is set for
MultiResolutionPDEDeformableRegistration using
SetNumberOfLevels(numLevels), and the iterations per level is
subsequently set using SetNumberOfIterations(itk::Array numIterations)
then numIterations.GetSize() overrides the earlier
SetNumberOfLevels(numLevels) command.  That is, the latter command
effectively executes SetNumberOfLevels(numIterations.GetSize()).

Add test for this functionality.

Note that SetNumberOfIterations(unsigned int *numIterations) does not
override the number of levels because levels information is not
available.