DOC: fixing "unbalanced grouping commands" warnings
Brute force approach to fix the "unbalanced grouping commands" warnings by explicitly placing the `ALIASES` `ITKStartGrouping` and `ITKEndGrouping` in the code where unbalanced grouping commands" warnings an attempt was made to insert the commands `@{` and `@}` and afterwards solving the warnings that were already present in respect to the "unbalanced grouping commands" warnings and also solving the warnings that appeared due to these changes.

The usage of the script has been abandoned as this implicitly added the commands based on the layout and this is is hard to enforce and it is also hard to get the script without errors (would nearly require a compiler).

This is related to issue #3654.