STYLE: Remove local `ostringstream loc` variable from ExceptionObject
The name `loc` was somewhat confusing here, as it is used elsewhere in
ExceptionObject as the parameter name for the `ITK_LOCATION` argument (the name
or signature of the function that threw the exception). Moreover, the use of a
string stream is no longer necessary, now that we can use `std::to_string`.