STYLE: Replace "Unknown" default exception location with nullptr or `{}`
It is preferable not duplicate a literal string so many times (following the DRY
principle). Moreover, it appears sufficient to use an empty value (`nullptr` or
`{}`), to indicate that the location of an exception is unspecified.

Note that an unspecified location may be "unknown", but instead, it may also be
"irrelevant", or "hidden". So it may depend on the application.

Triggered by a discussion with Bradley Lowekamp at
https://github.com/InsightSoftwareConsortium/ITK/pull/5266#discussion_r1985469317