STYLE: Remove Exception constructor overloads for `char *` arguments
Internally, `ExceptionObject` stores its file name, description, and location
as `std::string`, so the corresponding constructor parameters might as well just
be of type `std::string`.

Whenever such a parameter was originally declared `const std::string &`, this
commit has removed the `const` and the `&`, in order to efficiently "move" the
parameters from derived exception classes to their base class.

Note that this commit drops the support for null pointers as arguments.