ENH: Reset conj-grad learning rate for multi-start optimization
When used as a local optimizer in multi-start optimization,
itkConjugateGradientLineSearchOptimizerv4 gives inconsistent results because
its learningRate is carried over from the end of one optimization to the
beginning of the next.

Add an instance variable to hold the initial learning rate value, and
add the corresponding getter and setter macros.

Add a function to store the initial learning rate when SetLearningRate is
called - this is then restored on the first iteration when StartOptimization
is called.

Add a test to check that repeated optimizations with the same starting
parameters converge to the same result.