ENH: Reset conj-grad learning rate for multi-start optimization
When used as a local optimizer in multi-start optimization,
itkConjugateGradientLineSearchOptimizerv4 gives inconsistent results because
its learningRate is carried over from the end of one optimization to the
beginning of the next.

Added a function to store the initial learning rate when SetLearningRate is
called - this is then restored on the first iteration when StartOptimization
is called.

Added a test to check that repeated optimizations with the same starting
parameters converge to the same result.