DOC: Corrected setting for `CASE_SENSE_NAMES`
The default setting for `CASE_SENSE_NAMES` (in doxygen) is `SYSTEM` which means that on Linux it is set to `YES` and on (e.g.) Windows to `NO` (see analysis in https://github.com/InsightSoftwareConsortium/ITK/pull/5262#issuecomment-2765915282).
Seen the usage of the `EXCLUDE_PATTERN` `*/test*` this meant on Windows that also the directory with the name `TestKernel` was matched and thus excluded.
Setting the `CASE_SENSE_NAMES` explicitly to `YES` prevents this (on Linux it won't have any effect as here the default was sufficient).
Note: Due to a bug in doxygen (see https://github.com/doxygen/doxygen/pull/11519) this would still exclude the pattern. The new setting will only have an effect, on Windows, when, the not yet released, doxygen 1.14.0 or newer will be used.