BUG: Fix case where DICOM series order has negative spacing
If the spacing between slices is negative, then direction cosine
matrix was not correct with reguards to the sign.

This can be easily created with the "ReverseOrder" flag with the
series reader and DICOM data. With the correction, a DICOM series
mantains its correct physical location when this flag is enabled.