BUG: Jacobians not in physical space
The gradient is calculated in a neighborhood defined in index space.

DisplacementFieldTransform accounts for this by multiplying rows of the
gradient by the direction matrix. DisplacementFieldJacobianDeterminantFilter did not,
hence its determinants were different for the same physical transformation under different
voxel index ordering.

DisplacementFieldJacobianDeterminantFilter is now fixed to use the direction matrix.

The test is updated to check that the determinant for a given displacement field is consistent
at the same point in physical space, regardless of how the index space is oriented.

More details in issue #5358.