COMP: Update defaults for clang-tidy preferences
clang-tidy is integrated in several IDE's as an add on to
suggest code conformance to best practices standards.  The
existence of a .clang-tidy configuration file helps to
enforce a consistent set of conformance criteria.

These new updates provide assistance for keeping the codebase
clean with respect to community standards and best practice
recommendations.

NOTE: clang-tidy conformance is *not* enforced or checked
automatically.  This file provides consistency for those
who opt-in to using the tools.