STYLE: Prefer member initializer to assignment
Use member initializations in the constructor body which can be converted into
member initializers of the constructor instead. This not only improves the
readability of the code but also positively affects its performance.
Class-member assignments inside a control statement or following the first
control statement are ignored.

This check implements C.49 from the C++ Core Guidelines.

Please note, that this check does not enforce rule C.48 from the C++ Core
Guidelines. For that purpose see check modernize-use-default-member-init.

cppcoreguidelines-prefer-member-initializer