ENH: Fixed width type enums (INT8, UINT64, FLOAT64, ...) for IOComponent
Added enumerators to `IOComponent` that explicitly specify the number of bits
of the types they refer to:

    UINT8, INT8, UINT16, INT16, UINT32, INT32, UINT64, INT64, FLOAT32, FLOAT64

Tested that `MapPixelType<TPixel>::CType` maps to the expected enumerator.

`INT8` is clearer than `CHAR`, because the C++ type `char` may be either signed
or unsigned. (`INT8` corresponds with `std::int8_t`, which is always signed, by
definition.) `INT32` and `INT64` are clearer than `LONG`, because `long` may be
either 32 bits or 64 bits, depending on the platform and the compiler. `FLOAT32`
is clearer than `FLOAT`, because on other languages (notably Python), the
standard `float` type may be 64 bits.